url helpers refactoring
