+++ lib/pagy/extras/standalone.rb
@@ -40,7 +40,7 @@ class Pagy
       params[p_vars[:page_param].to_s]  = page
       params[p_vars[:items_param].to_s] = p_vars[:items] if defined?(ItemsExtra)
       # no Rack required in standalone mode
-      query_string = "?#{QueryUtils.build_nested_query(pagy_get_params(params))}" unless params.empty?
+      query_string = "?#{QueryUtils.build_nested_query(pagy_massage_params(params))}" unless params.empty?
       "#{url}#{query_string}#{p_vars[:fragment]}"
     end
   end
