+++ test/mock_helpers/controller.rb
@@ -6,7 +6,7 @@ class MockController
   include Pagy::Backend
   # we ned to explicitly include this because Pagy::Backend
   # does not include it when the test loads this module witout the headers
-  include Pagy::Helpers
+  include Pagy::UrlHelpers
 
   attr_reader :params
